update redux files
