+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Bookstore CMS",
+  "name": "Bookstore CMS",
   "icons": [
     {
       "src": "favicon.ico",
